Pulled in styling and layout from umberlas.com
I pulled in the style.css and experimental.css files from the
umberlas.com site. I also pulled in the layout from the index2.html file
of the umberlas.com site. I'm currently messing around with modifying
that general layout for a retuning of the fullrangefighting.com design.
Currently this is in the beginning stages. It still very much looks like
umberlas.com.